Mileage // Resolve bugs that were causing broken behavior
The new data file provided had an additional column that was not being accounted for
in Auto.new, therefore I added a new instance variable called `row_id` to capture
that first columns value.

Next, I added `.compact` to solve the nil comparison error in the sorting of the array
of mileage string values.

Lastly, I added a quality of life improvement to the argument in mileage.rb so that
even imperfect command line color entries would still work.